refactor(format machine data): moved printing specific utilities to a specific folder
