+++ src/common/utils/printing/formatRawMachineData.js
