+++ src/common/utils/printing/formatRawMachineData.test.js
